third_party/WebKit/LayoutTests/fast/forms/setrangetext-out-of-range-expected.txt

This tests the selection of the text field when mouse event occurs out of range

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5

the selection should be cleared
PASS textfield.selectionStart is non-zero.
PASS textfield.selectionEnd is non-zero.
PASS successfullyParsed is true

TEST COMPLETE
